Refactor : 알림 전송하는 회원 필드를 member -> sender로 수정
